requirements.txt
# Lobster Notes - Python Dependencies
# Database & Backend
flask==3.0.0
flask-cors==4.0.0
flask-mysqldb==2.0.0
mysqlclient==2.2.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
yt-dlp==2023.11.16

# Data Processing
lxml==4.9.3
soupsieve==2.5
